Refactor OrderItem entity and DAO to use Long for ID, enhance OrderItemService with error handling, and implement OrderItemController with CRUD operations and security configuration.

+++ orderTracking/src/main/java/com/verAuto/orderTracking/service/OrderItemService.java
@@ -9,7 +9,7 @@ public interface OrderItemService {
     List<OrderItem> findByCompanyName(String companyName);
     List<OrderItem> findByDestination(String destination);
     List<OrderItem> findByRegistrationNumber(String registrationNumber);
-    OrderItem findById(int id);
+    OrderItem findById(Long id);
     OrderItem save(OrderItem orderItem);
-    void deleteById(int id);
+    void deleteById(Long id);
 }
